feat: option to disable using remote registries when getting peerDeps lists
`--only-peers` was doing two different things which complicated the need
for proxy settings. Not only was it flagging whether to install the main
package with the peers, it was also toggling whether the peerDeps list
was being generated from a local node_modules vs a remote registry.

By splitting this into two distinct feature flags, we make behavior clearer,
and future case of delegating proxy settings to NPM or Yarn instead
of handling those in the utility.

BREAKING_CHANGE: `--only-peers` now searches for peerDeps via the registry
unless `--no-registry` is also specified

* was: `install-peerdeps <package> --only-peers`
* now: `install-peerdeps <package> --only-peers --no-registry`